Commit and push UserService.java and IUserService.java

+++ src/main/java/za/ac/cput/service/IUserService.java
@@ -3,8 +3,9 @@ package za.ac.cput.service;
 import za.ac.cput.domain.User;
 import java.util.List;
 
-public interface IUserService extends IService<User, Integer> {
-    boolean delete(Integer id);
+public interface IUserService extends IService<User, Long> {
+    boolean delete(Long id);
     List<User> getAll();
 }
+
